Address Improper calculate balance bug
